refactor(sync): extract bulk (all-projects) + branch-linking ops (tech-debt)
Continues the sync_service.py split. No behavior change.

- `_sync_bulk.py`: the all-projects orchestrators (pull_all, diff_all,
  push_all) as free functions taking the service; the public methods are now
  thin delegators.
- `_sync_branch.py`: git-branch <-> Keboola-dev-branch linking (branch_link,
  branch_unlink, branch_status). branch_link takes the service; the others are
  pure. `get_current_branch` is imported LOCALLY inside each function (as the
  original did) so tests that patch `keboola_agent_cli.sync.git_utils.
  get_current_branch` keep working -- a top-level import would capture the
  reference at import time and defeat source-module patching.

Public methods preserved as shims (commands + tests call them unchanged).
`sync_service.py`: 2581 lines (was 2931 this commit; 4033 at the PR base).
make check green (4092 passed).
